web/requirements_backend.txt
django
djangorestframework
numpy
xgboost
django-cors-headers
djangorestframework-simplejwt
scikit-learn
torch